refactor to use same standard, using TheoryData instead of Records
